add battery management files

+++ source/prototype.h
@@ -18,7 +18,8 @@ void MFET_set_up();
 void Select_mux_pin(int channel);
 
 float Read_Cell_Voltage(void);
-int Read_shunt_resistor(void);
+
+float Read_shunt_resistor(void);
 
 int Internal_Temp_Read();
